Setting headers for CORS, for preflight and authentication
Allowing requests from all origins. Once the request reaches the
authentication route, the methods that can be used to access the server
are decided based on the user being successfully authenticated i.e. if a
valid user auth token is not included in the request, only GET requests
will be accepted