Fixes to server, as per testing results
database_query.js - Adding create_Project to the module exports (was not
available outside this file)
validation.js - Ignoring type check if html or link, as these are not
dealt with in validation yet
feeds-actions.js - since the "previewCommit" action was moved from
feeds-read.js, it no longer has access to the project content. Removed
addition of structure to response, as this is not needed anyway
feeds-create.js - fix to issue caused during addition of documentation
comments - :itemPath made the request parameters limited to 2, whereas *
allows them to be any length. Also fixing the validation of content to
be added to properties, by passing the structure of that specific
property to the validation, as opposed to the structure of the items of
the object in which it exists
feeds-update.js - once entire project content file is updated, the
request was throwing an error, as it could not pass the next router of
its route. Swapping the order of the if statements, so that the request
can pass through. Also fixing issue with updating property of an array
item, by setting the structureFileData equal to the items of the array
(if they exist).